tttt

+++ Projecto_JavaBank/src/BancoAppAdm.java
@@ -355,8 +355,8 @@ public class BancoAppAdm implements Serializable {
 				int totalCapital = gb.javabank.totalCapital(gb.javabank.getContas(), data1, data2);
 				textFieldTotalCapital.setText(Integer.toString(totalCapital));
 
-				int balanco = gb.javabank.balanco(gb.javabank.getContas(), data1, data2);
-				textFieldBalanco.setText(Integer.toString(balanco));
+				double balanco = gb.javabank.balanco(gb.javabank.getContas(), data1, data2);
+				textFieldBalanco.setText(Double.toString(balanco));
 
 			}
 		});
